Updating source to Swift 1.2

+++ JelloSwift/Bitmask.swift
@@ -8,19 +8,19 @@
 
 import UIKit
 
-typealias Bitmask = Int;
+typealias Bitmask = UInt;
 
 infix operator +& { associativity left precedence 140 }
 infix operator -& { associativity left precedence 140 }
 
 func +&(inout lhs: Bitmask, rhs: Int) -> Bitmask
 {
-    lhs |= 1 << (rhs > 0 ? (rhs - 1) : 0);
+    lhs |= 1 << (UInt(rhs) > 0 ? (UInt(rhs) - 1) : 0);
     return lhs;
 }
 
 func -&(inout lhs: Bitmask, rhs: Int) -> Bitmask
 {
-    lhs &= ~(1 << (rhs > 0 ? (rhs - 1) : 0));
+    lhs &= ~(1 << (UInt(rhs) > 0 ? (UInt(rhs) - 1) : 0));
     return lhs;
 }
